Break ground on a prototype of the translation concept.
1. Draft translation of the first couple of chapters.
2. Rust project added under ports directory.

I'm not 100% on the best model for tracking upstream.

Upstream is written in a language tuple of (English/C++).

There are many potential downstream tuples. The first will
be (English/Rust).

I have a feeling that raytracing-translations to raytracing
should be as kubuntu is to unbuntu. As such, upstream
shouldn't have to bother with catering to downstream ports.

Initially, let's explore what a satisfactorily translated
document should look like.